jove.c: _uint_to_string: use 64-bit integer math, since this function may be provided with a u64

+++ lib/jove.util.c
@@ -64,7 +64,7 @@ static _INL void _uint_to_string(uint64_t x, char *Str, unsigned Radix) {
   char Buffer[65];
   char *BufPtr = &Buffer[sizeof(Buffer)];
 
-  uint32_t N = x;
+  uint64_t N = x;
 
   while (N) {
     *--BufPtr = Digits[N % Radix];
